Refactor CLI to reduce function complexity (especially for client paths)
