+++ src/client.rs
@@ -1,2 +1,3 @@
+pub mod cli;
 pub mod config;
 pub mod http;
